+++ changelog.md
@@ -11,6 +11,7 @@
 
 - [#4689](https://github.com/ignite/cli/pull/4689) Revert `HasGenesis` implementation from retracted `core` v1 to SDK `HasGenesis` interface.
 - [#4701](https://github.com/ignite/cli/pull/4701) Improve `ignite doctor` by removing manual migration step. Additionally, remove protoc to buf migrations logic.
+- [#4703](https://github.com/ignite/cli/pull/4703) Remove testutil.Register function.
 
 ### Bug Fixes