+++ changelog.md
@@ -12,6 +12,7 @@
 - [#4689](https://github.com/ignite/cli/pull/4689) Revert `HasGenesis` implementation from retracted `core` v1 to SDK `HasGenesis` interface.
 - [#4701](https://github.com/ignite/cli/pull/4701) Improve `ignite doctor` by removing manual migration step. Additionally, remove protoc to buf migrations logic.
 - [#4703](https://github.com/ignite/cli/pull/4703) Remove testutil.Register function.
+- [#4702](https://github.com/ignite/cli/pull/4702) Improve app detection by checking for inheritance instead of interface implementation.
 
 ### Bug Fixes