+++ changelog.md
@@ -16,8 +16,8 @@
 
 ### Bug Fixes
 
+- [#4691](https://github.com/ignite/cli/pull/4691), [#4706](https://github.com/ignite/cli/pull/4706) Fix ts-client query template and solely Go template for `ts-client` generation.
 - [#4686](https://github.com/ignite/cli/pull/4686) Filter discovered protos to only messages.
-- [#4691](https://github.com/ignite/cli/pull/4691), [](https://github.com/ignite/cli/pull/) Fix ts-client query template and solely Go template for `ts-client` generation.
 
 ## [`v29.0.0-rc.1`](https://github.com/ignite/cli/releases/tag/v29.0.0-rc.1)