+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [`v29.0.0`](https://github.com/ignite/cli/releases/tag/v29.0.0)
+
 ### Features
 
 - [#4683](https://github.com/ignite/cli/pull/4683) Allow to change default denom via flag.